Fix pooled costs during app to app recursion
Also adds divw, itxnas, gitxnas

Makes accessing 0 accounts in dry-run ok.

+++ data/transactions/logic/blackbox_test.go
@@ -87,7 +87,7 @@ func TestNewAppEvalParams(t *testing.T) {
 				if reflect.DeepEqual(param, config.Consensus[protocol.ConsensusV29]) {
 					require.Nil(t, ep.PooledApplicationBudget)
 				} else if reflect.DeepEqual(param, config.Consensus[protocol.ConsensusFuture]) {
-					require.Equal(t, *ep.PooledApplicationBudget, uint64(param.MaxAppProgramCost*testCase.numAppCalls))
+					require.Equal(t, *ep.PooledApplicationBudget, param.MaxAppProgramCost*testCase.numAppCalls)
 				}
 			})
 		}
